README.md
This is a simple fork of Peter Kim's The Hacker's Playbook 2 text file with the optional tools for installation in Kali Linux.  There was a deadlink and I wanted to make sure it worked on my machine.  I saved the resulting bash file as thp2_setup.sh.  Feel free to use it.  My fork is subject to Peter Kim's license terms for the original text file.  I am not adding any additional licensing terms and I waive any rights as an author to the simple changes incorporated in my bash file.